src/requirements.txt
requests==2.25.1
requests-oauthlib==1.3.0
python-dotenv==0.17.1
tweepy==3.10.0
typing-extensions==3.7.4.3
jsonschema==3.2.0
bs4==0.0.1